chore: move ol-contextmenu to dependencies and bump version
