+++ package.json
@@ -26,6 +26,7 @@
     "lint": "^1.1.2",
     "lodash": "^4.17.15",
     "ol": "^10.2.1",
+    "ol-contextmenu": "^5.5.0",
     "ol-ext": "^4.0.24",
     "proj4": "^2.4.4",
     "voca": "^1.4.0"
