+++ src/anol/anol.js
@@ -30,7 +30,7 @@ import DynamicGeoJSON from './layer/dynamicgeojson.js';
 import BBOXGeoJSON from './layer/bboxgeojson.js';
 import GMLLayer from './layer/gml.js';
 
-import ContextMenu from './contextmenu/base.js';
+import ContextMenu from './contextmenu/ContextMenu.js';
 import Control from './control.js';
 
 export var geocoder = {
