--- src/anol/contextmenu/README.rst
@@ -1,1 +1,0 @@
-ContextMenu Control based on ol-contextmenu from Jonatas Walker (https://github.com/jonataswalker/ol-contextmenu - MIT License) modified for anol. Thanks Jonatas for the great work.
